task: fully function based component
