+++ dataworkspace/dataworkspace/static/js/react/features/your-files/components/Modal/index.jsx
@@ -1,4 +1,5 @@
 import React, { useEffect, useRef } from 'react';
+
 import styled from 'styled-components';
 
 const Dialog = styled('dialog')`
