--- dataworkspace/dataworkspace/static/js/react/features/your-files/popups/index.js
@@ -1,3 +1,0 @@
-export { AddFolderPopup } from './AddFolder';
-export { UploadFilesPopup } from './UploadFiles';
-export { Modal } from './Modal';
