+++ dataworkspace/dataworkspace/static/js/react/features/your-files/styles/App.scss
@@ -283,9 +283,6 @@
   text-align: right;
   margin-bottom: 0px;
   max-width: 500px;
-  .modal-button {
-    margin-bottom: 0;
-  }
 }
 .modal-contents{
   overflow-y: scroll;
